requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
flake8==4.0.1
idna==3.3
mccabe==0.6.1
numpy==1.21.5
pandas==1.3.5
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
six==1.16.0
urllib3==1.26.7
